Revert "Foundation Set Up for Team Website Routing"
