+++ website/src/Pages/LandingPage/Components/SignInForm.module.css
@@ -38,7 +38,6 @@
     display: inline-block;
     border-bottom: 1.5px solid #FAC01A;
     padding-bottom: 5px;
-    cursor: pointer;
 }
 
 .formFieldInput {
