--- website/src/Pages/TeamWebsite/Components/Deliverables.js
@@ -1,65 +1,0 @@
-import { React } from "react";
-import Paper from '@mui/material/Paper';
-import Typography from "@mui/material/Typography";
-
-const Deliverables = () => {
-    return(
-        <div style={{height: '100vh', paddingLeft: '300px', backgroundColor: 'white', justifyContent: 'center', alignItems: 'center', display: 'flex', flexFlow: 'column'}}>
-            <Paper elevation={4} style={{height: '90%', width: '90%', margin: '10px', padding: '10px'}}>
-                <Typography
-                    variant="h4"
-                >
-                    Deliverable 1
-                </Typography>
-                <Typography
-                >
-                    Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.
-
-
-                </Typography>
-            </Paper>
-            <Paper elevation={4} style={{height: '90%', width: '90%', margin: '10px', padding: '10px'}}>
-            <Typography
-                    variant="h4"
-                >
-                    Deliverable 2
-                </Typography>
-                <Typography
-                >
-                    Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.
-
-
-                </Typography>
-            </Paper>
-            <Paper elevation={4} style={{height: '90%', width: '90%', margin: '10px', padding: '10px'}}>
-            <Typography
-                    variant="h4"
-                >
-                    Deliverable 3
-                </Typography>
-                <Typography
-                >
-                    Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.
-
-
-                </Typography>
-            </Paper>
-            <Paper elevation={4} style={{height: '90%', width: '90%', margin: '10px', padding: '10px'}}>
-            <Typography
-                    variant="h4"
-                >
-                    Deliverable 4
-                </Typography>
-                <Typography
-                >
-                    Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.
-
-
-                </Typography>
-            </Paper>
-            
-        </div>
-    )
-}
-
-export default Deliverables;
